pyproject.toml
[project]
name = "my-radio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]